Create IOs tests.
This involved moving tests for methods that were in Table.
